pyproject.toml
[tool.poetry]
name = "myprecious"
version = "0.1.0"
description = ""
authors = ["BiRabittoh <birabittoh@duck.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^2.3.2"
flask-sqlalchemy = "^3.0.5"
flask-login = "^0.6.2"
werkzeug = "^2.3.6"
wtforms = "^3.0.1"
flask-wtf = "^1.1.1"
igdb-api-v4 = "^0.2.0"
python-dotenv = "^1.0.0"
argon2-cffi = "^23.1.0"

[tool.poetry.group.prod]
optional = true

[tool.poetry.group.prod.dependencies]
waitress = "^2.1.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
